Remove mock-socket from prod deps
