+++ packages/rpc-provider/package.json
@@ -30,11 +30,11 @@
     "@polkadot/x-global": "^14.0.3",
     "@polkadot/x-ws": "^14.0.3",
     "eventemitter3": "^5.0.1",
-    "mock-socket": "^9.3.1",
     "tslib": "^2.8.1"
   },
   "devDependencies": {
     "@substrate/connect": "0.8.11",
+    "mock-socket": "^9.3.1",
     "nock": "^13.5.5"
   },
   "optionalDependencies": {
