+++ packages/rpc-provider/src/mock/types.ts
@@ -1,20 +1,6 @@
 // Copyright 2017-2026 @polkadot/rpc-provider authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Server } from 'mock-socket';
-
-export type Global = typeof globalThis & {
-  WebSocket: typeof WebSocket;
-  fetch: any;
-}
-
-export interface Mock {
-  body: Record<string, Record<string, unknown>>;
-  requests: number;
-  server: Server;
-  done: () => Promise<void>;
-}
-
 export type MockStateSubscriptionCallback = (error: Error | null, value: any) => void;
 
 export interface MockStateSubscription {
