+++ packages/rpc-provider/src/ws/connect.spec.ts
@@ -3,9 +3,9 @@
 
 /// <reference types="@polkadot/dev-test/globals.d.ts" />
 
-import type { Mock } from '../mock/types.js';
+import type { Mock } from '../mock/testTypes.manual.js';
 
-import { mockWs } from '../mock/mockWs.js';
+import { mockWs } from '../mock/mockWs.manual.js';
 import { WsProvider } from './index.js';
 
 const TEST_WS_URL = 'ws://localhost-connect.spec.ts:9988';
